Changing verification text to reflect changes in thought
So this has a couple of edits, first it changes the verification selection
process such that we chose routes we are *not* using because it's been
generally agreed that routes we are using may as well be verified continuously.

This also changes some of the wording of the selection method to be more
rigorous by stating that we intend a simple random sample. As opposed to other
possible sampling methods, some of which may even have advantages for us if
significant effort was invested in utilizing them well (clustered samples for
example could help narrow down sources of inaccuracy).

As a final note it might be useful to look into how many routes a given node
must sample for a given routing table size in order for every route advertised
to be verified within some reasonable period and confidence interval.